funcion editar para todos los posts

+++ src/lib/showPost.js
@@ -16,26 +16,25 @@ export const createPost = (posts) => {
       <textarea class="post-content" rows="4" cols="50" readonly>${iterator[0].data.content}</textarea>
       <textarea class="reference-link" readonly>${iterator[0].data.referenceLink}</textarea>
       </div>`;
-
     containerPost.innerHTML += templatePost;
   };
   posts.forEach(postContent);
   const deleteBtn = containerPost.querySelectorAll('.delete-btn');
   deleteBtn.forEach((btn) => {
     btn.addEventListener('click', () => {
-      deleteDocData(btn.value);
+      if(confirm('¡Vas a borrar tu post! ¿Deseas continuar?')) {
+        deleteDocData(btn.value);
+      }
     });
   });
   const editBtn = containerPost.querySelectorAll('.edit-btn');
-  const headerPost = containerPost.querySelector('.header-post');
-  const contentPost = containerPost.querySelector('.post-content');
-  const linkReference = containerPost.querySelector('.reference-link');
 
   editBtn.forEach((btn) => {
-    btn.addEventListener('click', () => {
-      headerPost.removeAttribute('readonly');
-      contentPost.removeAttribute('readonly');
-      linkReference.removeAttribute('readonly');
+    btn.addEventListener('click', (event) => {
+      const targetBtn = event.target;
+      targetBtn.previousElementSibling.removeAttribute('readonly');
+      targetBtn.parentElement.nextElementSibling.removeAttribute('readonly');
+      targetBtn.parentElement.nextElementSibling.nextElementSibling.removeAttribute('readonly');
       const postId = btn.value;
       const parentDivPost = document.getElementById(postId);
       const header = parentDivPost.querySelector('.header-post').value;
